feat(Thread): Enable old thread controls
