+++ src/modules/Thread/components/Controllers/CommentController.jsx
@@ -5,26 +5,26 @@ export default function(props) {
     // const { } = props;
 
     const toggledProps = {
-        name: "comment-text"
+        name: "mesage-1"
     }
 
     const defaultProps = {
-        name: "comment-text"
+        name: "mesage-1"
     }
 
     {/* OR
         right facing:
         <Icon name="message" />
         <Icon name="message-text" />
-        
+
         left facing:
-        <Icon name="message-reply-text" />  
+        <Icon name="message-reply-text" />
         <Icon name="message-reply-text" />
         <Icon name="reply" />  addow facing left
 
     */}
 
     return <ButtonCircle toggleProps={toggledProps}>
-        <Icon {...defaultProps} /> 
+        <Icon {...defaultProps} />
     </ButtonCircle>
 }
