+++ src/modules/Thread/containers/ThreadControls/ThreadControls.styles.sass
@@ -10,7 +10,14 @@ $thread-controls-height: 65px
     left: 0
     z-index: 20
     height: 1px
-    transform: translateY($thread-controls-height)
+
+    &.animate-in
+        transform: translateY(0)
+        transition: transform .65s ease
+
+    &.animate-out
+        transform: translateY($thread-controls-height)
+        transition: transform .2s ease-out
 
     .left-controls
         left: 0
@@ -28,10 +35,3 @@ $thread-controls-height: 65px
         height: $thread-controls-height
         bottom: 0
         padding: 0 $thread-controls-gutter-size
-
-    &.animate-in
-        animation: moveUp .65s ease forwards .4s
-
-    &.animate-out
-        animation: moveUp .2s ease forwards
-        animation-direction: reverse
